TW-2301: split evm fetch module, fix SIRS activity filter
