+++ src/activity/evm/fetch.ts
@@ -1,41 +1,16 @@
-import { zeroAddress } from 'viem';
-
-import {
-  EtherlinkAccountTransactionsPageParams,
-  EtherlinkCoinBalanceHistoryItem,
-  EtherlinkInternalTransaction,
-  EtherlinkInternalTransactionsPageParams,
-  EtherlinkLog,
-  EtherlinkTokenTransfer,
-  EtherlinkTokenTransfersPageParams,
-  EtherlinkTransaction,
-  EtherlinkTransactionLogsPageParams,
-  fetchGetAccountCoinBalanceHistory,
-  fetchGetAccountTokenTransfers,
-  fetchGetAccountTransactions,
-  fetchGetTransactionInternalTransactions,
-  fetchGetTransactionLogs,
-  fetchGetTransactionTokenTransfers,
-  isErc20TokenTransfer,
-  isErc721TokenTransfer
-} from 'src/apis/etherlink';
+import { EtherlinkTokenTransfer, EtherlinkTransaction } from 'src/apis/etherlink';
 import { TempleChainKind } from 'src/enums/temple-chain-kind.enum';
-import { DEFAULT_EVM_CURRENCY } from 'src/token/interfaces/token-metadata.interface';
-import { DEFAULT_EVM_CHAINS_SPECS } from 'src/types/networks';
 import { putToCappedCache } from 'src/utils/capped-cache.utils';
 import { equalsIgnoreCase } from 'src/utils/evm/on-chain/common.utils';
 import { isDefined } from 'src/utils/is-defined';
 
-import { ActivityOperKindEnum, ActivityOperTransferType, ActivityStatus, EvmActivity, EvmOperation } from '../types';
+import { ActivityStatus, EvmActivity } from '../types';
 import { throwIfAborted } from '../utils';
 
-import { EvmOperationKind, getEvmOperationKind } from './operation-kind';
-import { getApprovalLogOwnerAddress, isApprovalLog, parseApprovalLog } from './parse-approval';
+import { EtherlinkActivitiesPageParams, getEtherlinkHistoryData } from './fetchers';
+import { parseTokenTransfer, toBigInt, toUnorderedOperations } from './parse';
 
-export interface EtherlinkActivitiesPageParams {
-  operationsPageParams: EtherlinkAccountTransactionsPageParams | nullish;
-  tokensTransfersPageParams: EtherlinkTokenTransfersPageParams | nullish;
-}
+export type { EtherlinkActivitiesPageParams } from './fetchers';
 
 interface EtherlinkActivitiesPage {
   activities: EvmActivity[];
@@ -69,377 +44,6 @@ const putCachedActivity = (
     activity,
     MAX_CACHED_ACTIVITIES
   );
-
-const INTEGER_REGEX = /^-?\d+$/;
-
-const toBigInt = (value: string) => (INTEGER_REGEX.test(value) ? BigInt(value) : 0n);
-
-const fetchAllPages = async <T, P extends object>(
-  fetchPage: (pageParams: P | undefined, signal?: AbortSignal) => Promise<{ items: T[]; next_page_params: P | null }>,
-  signal?: AbortSignal
-) => {
-  const items: T[] = [];
-  let pageParams: P | undefined;
-
-  do {
-    throwIfAborted(signal);
-    const page = await fetchPage(pageParams, signal);
-    items.push(...page.items);
-    pageParams = page.next_page_params ?? undefined;
-  } while (isDefined(pageParams));
-
-  return items;
-};
-
-const fetchAllTxLogs = (txHash: string, signal?: AbortSignal) =>
-  fetchAllPages<EtherlinkLog, EtherlinkTransactionLogsPageParams>(
-    (pageParams, pageSignal) => fetchGetTransactionLogs(txHash, pageParams, pageSignal),
-    signal
-  );
-
-const fetchAllTxInternalTransactions = (txHash: string, signal?: AbortSignal) =>
-  fetchAllPages<EtherlinkInternalTransaction, EtherlinkInternalTransactionsPageParams>(
-    (pageParams, pageSignal) => fetchGetTransactionInternalTransactions(txHash, pageParams, pageSignal),
-    signal
-  );
-
-const fetchAllTxTokenTransfers = (txHash: string, signal?: AbortSignal) =>
-  fetchAllPages<EtherlinkTokenTransfer, EtherlinkTokenTransfersPageParams>(
-    (pageParams, pageSignal) => fetchGetTransactionTokenTransfers(txHash, pageParams, pageSignal),
-    signal
-  );
-
-const makeGasAsset = (chainId: number) => {
-  const currency = DEFAULT_EVM_CHAINS_SPECS.find(specs => specs.chainId === chainId)?.currency ?? DEFAULT_EVM_CURRENCY;
-
-  return { contract: currency.address, decimals: currency.decimals, nft: false, symbol: currency.symbol };
-};
-
-const makeGasTokenTransfer = (
-  operation: Pick<EtherlinkTransaction, 'from' | 'to' | 'value'> & { logIndex: number },
-  chainId: number,
-  accountAddress: string
-): EvmOperation => {
-  const { from, to, value, logIndex } = operation;
-  const isSending = equalsIgnoreCase(from.hash, accountAddress);
-
-  return {
-    kind: ActivityOperKindEnum.transfer,
-    fromAddress: from.hash,
-    toAddress: to?.hash ?? zeroAddress,
-    asset: { ...makeGasAsset(chainId), amountSigned: isSending ? `-${value}` : value },
-    logIndex,
-    type: isSending ? ActivityOperTransferType.sendToAccount : ActivityOperTransferType.receiveFromAccount
-  };
-};
-
-const parseTokenTransfer = (transfer: EtherlinkTokenTransfer, accountAddress: string): EvmOperation => {
-  const { from, to, log_index: logIndex } = transfer;
-  const isSending = equalsIgnoreCase(from.hash, accountAddress);
-  const amountNotSigned = isErc721TokenTransfer(transfer) ? '1' : transfer.total.value;
-  const amountSigned = isSending ? `-${amountNotSigned}` : amountNotSigned;
-
-  if (isErc20TokenTransfer(transfer)) {
-    const { address_hash: address, decimals, symbol, icon_url: iconURL } = transfer.token;
-
-    return {
-      kind: ActivityOperKindEnum.transfer,
-      fromAddress: from.hash,
-      toAddress: to.hash,
-      asset: {
-        contract: address,
-        amountSigned,
-        decimals: decimals == null ? undefined : Number(decimals),
-        nft: false,
-        symbol: symbol ?? undefined,
-        iconURL: iconURL ?? undefined
-      },
-      logIndex,
-      type: isSending ? ActivityOperTransferType.sendToAccount : ActivityOperTransferType.receiveFromAccount
-    };
-  }
-
-  const { total } = transfer;
-  const instance = total.token_instance;
-  const {
-    address_hash: address,
-    symbol,
-    decimals: rawDecimals,
-    icon_url: iconFallback
-  } = instance?.token ?? transfer.token;
-
-  return {
-    kind: ActivityOperKindEnum.transfer,
-    fromAddress: from.hash,
-    toAddress: to.hash,
-    asset: {
-      contract: address,
-      tokenId: instance?.id ?? total.token_id,
-      amountSigned,
-      decimals: rawDecimals == null ? 0 : Number(rawDecimals),
-      nft: true,
-      symbol: symbol ?? undefined,
-      iconURL: instance?.metadata?.image ?? iconFallback ?? undefined
-    },
-    logIndex,
-    type: isSending ? ActivityOperTransferType.sendToAccount : ActivityOperTransferType.receiveFromAccount
-  };
-};
-
-const getApprovalsForAccount = async (tx: EtherlinkTransaction, accountAddress: string, signal?: AbortSignal) => {
-  const logEntries = await fetchAllTxLogs(tx.hash, signal);
-
-  return logEntries
-    .map(({ topics, index, data, address, decoded }) => ({
-      topics: topics.filter(isDefined),
-      logIndex: index,
-      data,
-      address: address.hash,
-      decoded
-    }))
-    .filter(
-      ({ decoded, topics }) =>
-        isApprovalLog({ decoded, topics }) && equalsIgnoreCase(getApprovalLogOwnerAddress(topics), accountAddress)
-    )
-    .map(({ topics, logIndex, data, address }) => parseApprovalLog({ topics, logIndex, data, address }));
-};
-
-const toUnorderedOperations = async (
-  tx: EtherlinkTransaction,
-  tokensTransfers: EtherlinkTokenTransfer[],
-  nativeCoinDelta: string,
-  accountAddress: string,
-  chainId: number,
-  signal?: AbortSignal
-): Promise<EvmOperation[]> => {
-  const { to, position, raw_input: rawInput, value, status } = tx;
-  const toAddress = to?.hash;
-  const parsedTokensTransfers = tokensTransfers.map(transfer => parseTokenTransfer(transfer, accountAddress));
-
-  const operationKind = getEvmOperationKind({ data: rawInput, to: toAddress, value: toBigInt(value) });
-  const gasTokenTransfer = makeGasTokenTransfer({ ...tx, logIndex: position }, chainId, accountAddress);
-  const fallbackOperations: EvmOperation[] = [
-    {
-      kind: ActivityOperKindEnum.interaction,
-      logIndex: position,
-      withAddress: toAddress,
-      asset: Number(value) > 0 ? { ...makeGasAsset(chainId), amountSigned: `-${value}` } : undefined
-    }
-  ];
-
-  switch (operationKind) {
-    case EvmOperationKind.DeployContract:
-      return parsedTokensTransfers.concat(fallbackOperations);
-    case EvmOperationKind.Send:
-    case EvmOperationKind.Mint:
-      if (rawInput === '0x') {
-        return [gasTokenTransfer];
-      }
-
-      return parsedTokensTransfers.length
-        ? parsedTokensTransfers.concat(Number(value) ? gasTokenTransfer : [])
-        : fallbackOperations;
-    case EvmOperationKind.Approval: {
-      const approvals = await getApprovalsForAccount(tx, accountAddress, signal);
-
-      return approvals.length ? approvals : fallbackOperations;
-    }
-    case EvmOperationKind.ApprovalForAll:
-      return [{ kind: ActivityOperKindEnum.interaction, logIndex: position, withAddress: toAddress }];
-    default: {
-      if (status !== 'ok') {
-        return fallbackOperations;
-      }
-
-      const hasGasTokenReceiveOperations = toBigInt(nativeCoinDelta) + toBigInt(value) > 0n;
-
-      if (!tokensTransfers.length && !hasGasTokenReceiveOperations) {
-        const approvalOperations = await getApprovalsForAccount(tx, accountAddress, signal);
-
-        return approvalOperations.length ? approvalOperations : fallbackOperations;
-      }
-
-      let gasTokenReceiveOperations: EvmOperation[] = [];
-
-      if (hasGasTokenReceiveOperations) {
-        const internalOperations = await fetchAllTxInternalTransactions(tx.hash, signal);
-        gasTokenReceiveOperations = internalOperations
-          .filter(operation => equalsIgnoreCase(operation.to?.hash, accountAddress) && Number(operation.value) > 0)
-          .map(operation => makeGasTokenTransfer({ ...operation, logIndex: operation.index }, chainId, accountAddress));
-      }
-
-      return parsedTokensTransfers.concat(gasTokenReceiveOperations).concat(Number(value) ? gasTokenTransfer : []);
-    }
-  }
-};
-
-interface AlignedEtherlinkPage {
-  explicitOperations: EtherlinkTransaction[];
-  explicitOperationsNextPageParams: EtherlinkAccountTransactionsPageParams | nullish;
-  coinBalanceHistoryItems: EtherlinkCoinBalanceHistoryItem[];
-  tokensTransfers: EtherlinkTokenTransfer[];
-  tokensTransfersNextPageParams: EtherlinkTokenTransfersPageParams | nullish;
-}
-
-interface AlignEtherlinkPagesInput {
-  accountAddress: string;
-  explicitOperations: EtherlinkTransaction[];
-  explicitOperationsNextPageParams: EtherlinkAccountTransactionsPageParams;
-  coinBalanceHistoryItems: EtherlinkCoinBalanceHistoryItem[];
-  tokensTransfers: EtherlinkTokenTransfer[];
-  tokensTransfersNextPageParams: EtherlinkTokenTransfersPageParams;
-  operationsPageParams: EtherlinkAccountTransactionsPageParams | nullish;
-  tokensTransfersPageParams: EtherlinkTokenTransfersPageParams | nullish;
-  /** All transfers of the boundary transaction, pre-fetched when the transfers page ends above the operations page */
-  lastTxTokensTransfers: EtherlinkTokenTransfer[] | undefined;
-}
-
-const alignEtherlinkPages = (input: AlignEtherlinkPagesInput): AlignedEtherlinkPage => {
-  const { accountAddress, operationsPageParams, tokensTransfersPageParams, lastTxTokensTransfers } = input;
-  let { explicitOperations, coinBalanceHistoryItems, tokensTransfers } = input;
-  let explicitOperationsNextPageParams: EtherlinkAccountTransactionsPageParams | nullish =
-    input.explicitOperationsNextPageParams;
-  let tokensTransfersNextPageParams: EtherlinkTokenTransfersPageParams | nullish = input.tokensTransfersNextPageParams;
-
-  const lastTransfer = tokensTransfers.at(-1);
-  const lastOperation = explicitOperations.at(-1);
-
-  if (lastTxTokensTransfers && lastTransfer) {
-    const lastTransferHash = lastTransfer.transaction_hash;
-    tokensTransfers = tokensTransfers
-      .filter(({ transaction_hash: txHash }) => txHash !== lastTransferHash)
-      .concat(
-        lastTxTokensTransfers
-          .filter(({ from, to }) => [from, to].some(({ hash }) => equalsIgnoreCase(hash, accountAddress)))
-          .sort(({ log_index: aLogIndex }, { log_index: bLogIndex }) => bLogIndex - aLogIndex)
-      );
-
-    const lastRealignedTransfer = tokensTransfers.at(-1);
-
-    if (lastRealignedTransfer) {
-      const boundaryBlockNumber = lastRealignedTransfer.block_number;
-      tokensTransfersNextPageParams = { block_number: boundaryBlockNumber, index: lastRealignedTransfer.log_index };
-
-      explicitOperations = explicitOperations.filter(
-        ({ block_number: blockNumber }) => blockNumber >= boundaryBlockNumber
-      );
-      coinBalanceHistoryItems = coinBalanceHistoryItems.filter(
-        ({ block_number: blockNumber }) => blockNumber >= boundaryBlockNumber
-      );
-
-      const lastRemainingOperation = explicitOperations.at(-1);
-      explicitOperationsNextPageParams = lastRemainingOperation
-        ? {
-            block_number: lastRemainingOperation.block_number,
-            fee: lastRemainingOperation.fee?.value ?? '0',
-            hash: lastRemainingOperation.hash,
-            index: lastRemainingOperation.position,
-            inserted_at: lastRemainingOperation.timestamp,
-            // Only grows, never resets: the endpoint's own cursors count all items seen so far
-            items_count: (operationsPageParams?.items_count ?? 0) + explicitOperations.length,
-            value: lastRemainingOperation.value
-          }
-        : // Reuse the previous cursor (page 1 has none) - the next call retries this page while the other endpoint advances
-          operationsPageParams;
-    }
-  } else if (
-    input.explicitOperationsNextPageParams.block_number > input.tokensTransfersNextPageParams.block_number &&
-    lastOperation
-  ) {
-    const lastOperationBlockNumber = lastOperation.block_number;
-    const earliestBlockNumberOperationsHashes = new Set(
-      explicitOperations
-        .filter(({ block_number: blockNumber }) => blockNumber === lastOperationBlockNumber)
-        .map(({ hash }) => hash.toLowerCase())
-    );
-    tokensTransfers = tokensTransfers.filter(
-      ({ block_number: blockNumber, transaction_hash: txHash }) =>
-        blockNumber > lastOperationBlockNumber || earliestBlockNumberOperationsHashes.has(txHash.toLowerCase())
-    );
-
-    const lastRemainingTransfer = tokensTransfers.at(-1);
-    tokensTransfersNextPageParams = lastRemainingTransfer
-      ? { block_number: lastRemainingTransfer.block_number, index: lastRemainingTransfer.log_index }
-      : tokensTransfersPageParams;
-  }
-
-  return {
-    explicitOperations,
-    explicitOperationsNextPageParams,
-    coinBalanceHistoryItems,
-    tokensTransfers,
-    tokensTransfersNextPageParams
-  };
-};
-
-const getEtherlinkHistoryData = async (
-  currentOlderThan: EtherlinkActivitiesPageParams | undefined,
-  accountAddress: string,
-  signal?: AbortSignal
-): Promise<AlignedEtherlinkPage> => {
-  const { operationsPageParams, tokensTransfersPageParams } = currentOlderThan ?? {};
-
-  let explicitOperations: EtherlinkTransaction[] = [];
-  let explicitOperationsNextPageParams: EtherlinkAccountTransactionsPageParams | nullish = null;
-  let coinBalanceHistoryItems: EtherlinkCoinBalanceHistoryItem[] = [];
-
-  if (operationsPageParams !== null) {
-    throwIfAborted(signal);
-    const operationsPage = await fetchGetAccountTransactions(accountAddress, operationsPageParams, signal);
-    explicitOperations = operationsPage.items;
-    explicitOperationsNextPageParams = operationsPage.next_page_params;
-
-    throwIfAborted(signal);
-    const coinBalanceHistoryPage = await fetchGetAccountCoinBalanceHistory(
-      accountAddress,
-      operationsPageParams && {
-        block_number: operationsPageParams.block_number,
-        items_count: operationsPageParams.items_count
-      },
-      signal
-    );
-    coinBalanceHistoryItems = coinBalanceHistoryPage.items;
-  }
-
-  let tokensTransfers: EtherlinkTokenTransfer[] = [];
-  let tokensTransfersNextPageParams: EtherlinkTokenTransfersPageParams | nullish = null;
-
-  if (tokensTransfersPageParams !== null) {
-    throwIfAborted(signal);
-    const tokensTransfersPage = await fetchGetAccountTokenTransfers(accountAddress, tokensTransfersPageParams, signal);
-    tokensTransfers = tokensTransfersPage.items;
-    tokensTransfersNextPageParams = tokensTransfersPage.next_page_params;
-  }
-
-  const lastTransfer = tokensTransfers.at(-1);
-
-  if (!explicitOperationsNextPageParams || !tokensTransfersNextPageParams) {
-    return {
-      explicitOperations,
-      explicitOperationsNextPageParams,
-      coinBalanceHistoryItems,
-      tokensTransfers,
-      tokensTransfersNextPageParams
-    };
-  }
-
-  const lastTxTokensTransfers =
-    explicitOperationsNextPageParams.block_number <= tokensTransfersNextPageParams.block_number && lastTransfer
-      ? await fetchAllTxTokenTransfers(lastTransfer.transaction_hash, signal)
-      : undefined;
-
-  return alignEtherlinkPages({
-    accountAddress,
-    explicitOperations,
-    explicitOperationsNextPageParams,
-    coinBalanceHistoryItems,
-    tokensTransfers,
-    tokensTransfersNextPageParams,
-    operationsPageParams,
-    tokensTransfersPageParams,
-    lastTxTokensTransfers
-  });
-};
 
 const getOldestTimestamp = (items: { timestamp: string }[]) =>
   items.reduce<number | null>((oldest, { timestamp }) => {
